+++ CHANGELOG.md
@@ -6,3 +6,12 @@
 
 > ⚠️ The previous structure was compatible with early versions of Moodle 4.1 (pre-stable).  
 > This update ensures full compatibility with Moodle 5.x.
+
+------------
+
+## [2025042000] - 2025-04-20
+
+### Changed
+- Add two class in section: selector and navigation.
+- Changed default configs.
+- Added cache to save the last section where the user was.